Add update_tags (Rust binary)
This application represented an atomic chunk of work that formed a
good start to migrating the docker-volume-manager over to Rust
(where it will become more generic, useful, and be renamed to the
ContainerVolumeManager). Right now, the only purpose of this
application is to read the `tags' file associated with a volume
image and update it with the latest tag as a result of a backup
operation.