Build images w/ Containerfiles, integrate update_tags
Previously, images were built using a shell script that invoked
Buildah directly on each operation. This became cumbersome to
maintain, and it didn't allow caching of intermediate work. Refactor
the repository to build the images with Containerfiles and buildah's
bud command.

Also add a command to edtwardy-webservices to invoke the backup
function of the docker-volume-manager, and integrate the update_tags
program.