Fix bug that prevented the systemd service from restarting
systemd is smart enough to restart a service that has just been
reinstalled, whcih was not clear to me when I wrote the reinstall
target of the Makefile. Explicitly restarting the service after
reinstalling it would cause the service to stop while executing the
container volume manager, but surprisingly, this container would not
be removed by Docker, and thus an error would occur when next
starting the service (name already taken by an existing container).