Add ldap auth, better debug support
This patch primarily adds django-ldap-auth, and hooks the apps
image (Django) up to the openldap server for authentication.
Authentication is tested, and the volumes have been migrated. Also
add the apps_populator.sh script, which is a handy little Bash
script to start up a debug container that can be used to talk to
other containers (or more likely) to migrate/collectstatic files
into the volumes.