Fix regexp bug in docker-volume-manager
Previously, if the name of a volume was also a substring of its
tag, and there was another volume also versioned to that same tag,
docker-volume-manager would fail to compare the tags in the lockfile
and the tags file correctly (because awk would return a multiline
string for the match). Fix has been pushed to Dockerhub.