Fix renew-certificates.bash (AGAIN)
The script, which is currently configured to run daily, was not
detecting invalid certificates in the webroot. Apparently, at some
point, a certificate request had taken place for the domain
edtwardy.hopto.org from the server while a valid certificate was in
place, which caused certbot to emplace new certificates in a dir
with suffix -0001. Nginx was not updated to point to these new
certificates, and so when the served certs expired, the cron job
would see the -0001 certs and refuse renewal. All is well.

TODO: May have to update this script again in February to get
timezones working.