Add new synchronization point: services-up.target
This systemd target marks the point at which all backend services
are up and running. At this point, we can start the reverse proxy.
In practice, this doesn't work as well as I hoped, because the
starting of a systemd unit does not correspond with the moment that
it is accepting HTTP connections, and this has caused nginx to crash
on more than one occasion. So, we change the container@.service unit
to restart automatically. In the future, we will write healthcheck
methods that attempt to open TCP connections to backend services and
make use of sd_notify to synchronize the start of nginx.