nginx: Replace volumetric with squashfuse for siteconf
Previously, the siteconf volume was managed using volumetric. Now,
we are mounting it using squashfuse and the local volume driver. This
prevents a startup failure that can be observed if the
twardyece_nginx_1 container is not removed before the server finishes
shutting down, and volumetric attempts to freshen the volume contents
when the machine starts back up.