Replace blog and repository volumes with sitedata volume
This is a bit of a workaround until libpod v4.8.1 is packaged for Debian
and I can use Quadlet drop-in files, but it works for now. When drop-in
files become available, I will refactor these packages so that these
volumes are installed by the site packages and added to the nginx
container.