all: Fix various reproducibility bugs, update jenkins
I migrated to a new storage medium and installed all of these services
from scratch. Many of these changes had been made to
non-version-controlled files on the old system, or were somehow not
impacting the old system. Rebuilding the Jenkins containers triggered an
upgrade of the base Jenkins controller and agent versions, which
required some necessary configuration changes to allow the agent to
connect to the controller again.